+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 * Implementa a listagem de veículos dispoíveis na abertura de um socorro em rota (PL-2446)
 * Implementa a listagem de opções de problema disponíveis na abertura de um socorro em rota (PL-2447)
 * Implementa a listagem de socorros em rota (PL-2424)
+* Implementa a visualização de um socorro em rota (PL-2425)
 * Implementa o atendimento de socorros em rota (PL-2426)
 * Implementa a invalidação de socorros em rota (PL-2427)
 * Implementa a finalização de socorros em rota (PL-2428)